+++ app/build.gradle.kts
@@ -37,6 +37,7 @@ android {
         buildConfig = true
         viewBinding = true
         dataBinding = true
+        mlModelBinding = true
     }
 }
 
@@ -47,6 +48,7 @@ dependencies {
     implementation(libs.material)
     implementation(libs.androidx.activity)
     implementation(libs.androidx.constraintlayout)
+    implementation(libs.tensorflow.lite.gpu)
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
@@ -55,6 +57,15 @@ dependencies {
     implementation(libs.androidx.camera.camera2)
     implementation(libs.camera.lifecycle)
     implementation(libs.camera.view)
+
+    //Tensorflowlite
+    implementation(libs.tensorflow.lite.support)
+    implementation(libs.tensorflow.tensorflow.lite.metadata)
+    implementation(libs.tensorflow.tensorflow.lite.task.vision)
+
+    implementation(libs.circleimageview)
+    implementation(libs.glide)
+    implementation(libs.shimmer)
