+++ app/src/main/AndroidManifest.xml
@@ -6,6 +6,7 @@
         android:name="android.permission.WRITE_EXTERNAL_STORAGE"
         tools:ignore="ScopedStorage" />
     <uses-permission android:name="android.permission.CAMERA" />
+
     <uses-feature android:name="android.hardware.camera.any" />
 
     <application
@@ -18,6 +19,9 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Deras"
         tools:targetApi="31">
+        <activity
+            android:name=".ui.result.RresultActivity"
+            android:exported="false" />
         <activity
             android:name=".ui.profile.ProfileActivity"
             android:exported="false" />
@@ -39,9 +43,11 @@
         <activity
             android:name=".ui.login.LoginActivity"
             android:exported="false" />
-        <!--<activity
+        <!--
+<activity
             android:name=".ui.main.MainActivity"
-            android:exported="false" />-->
+            android:exported="false" />
+        -->
         <activity
             android:name=".ui.main.MainActivity"
             android:exported="true">
