--- app/src/main/java/com/deras/id/ui/customview/PasswordCV.kt
@@ -1,29 +1,0 @@
-package com.deras.id.ui.customview
-
-import android.content.Context
-import android.text.Editable
-import android.text.TextWatcher
-import android.util.AttributeSet
-import com.deras.id.R
-import com.google.android.material.textfield.TextInputEditText
-
-class PasswordCV @JvmOverloads constructor(
-    context: Context, attrs: AttributeSet? = null
-) : TextInputEditText(context, attrs) {
-
-    init {
-        addTextChangedListener(object : TextWatcher {
-            override fun beforeTextChanged(s: CharSequence?, start: Int, count: Int, after: Int) {}
-
-            override fun onTextChanged(s: CharSequence?, start: Int, before: Int, count: Int) {
-                if (s != null && s.length < 8) {
-                    setError(context.getString(R.string.passwords_must_have_at_least_8_characters), null)
-                } else {
-                    error = null
-                }
-            }
-
-            override fun afterTextChanged(s: Editable?) {}
-        })
-    }
-}
