+++ app/src/main/res/layout/activity_detection.xml
@@ -19,29 +19,18 @@
         app:layout_constraintTop_toTopOf="parent"
         tools:srcCompat="@tools:sample/avatars" />
 
-    <TextView
-        android:id="@+id/label_article"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="16dp"
-        android:text="@string/Result"
-        android:textColor="@color/deras_secondary_700"
-        android:textSize="16sp"
-        android:textStyle="bold"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/detection_image" />
-
     <Button
         android:id="@+id/btn_upload"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="16dp"
+        android:layout_marginTop="332dp"
         android:background="@drawable/custom_logout_button"
         android:text="@string/button_text_analiyze"
         android:textAllCaps="false"
         app:backgroundTint="@color/deras_secondary_500"
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/label_article" />
+        app:layout_constraintTop_toBottomOf="@+id/detection_image" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
